core, refactor 'ExpandBoxHack'
-move to common 'qltk.x' path
-rename to more appropriate 'ExpanderTitleContainerHack'
-add '__init__' function to populate content
